Added 'Psychotic Driving REMASTERED'
This is probably the greatest game I've made so far. Much like Postal 2 or 4, there is a lot of dark/offensive humor, so please do keep in mind that this game is a joke. It was made to be funny. If any of the stereotypes (furries, emos, vegans) offend you, this was originally made in primary school. This 2025 remaster simply adds QoL features and removes weird unnecessary features like a variable for how much fuel you've used. I'm sure the variable is still there and I simply hid it, but I can't be bothered checking. This is one of my bigger projects, at least in terms of code, because in visual scripting (block coding) the density and complexity of your code builds up over time, especially in big if-elses that I use for random sounds (because using the sound number is annoying, it means you have to organize your sounds so you can easily break the project by accidentally dragging and dropping the sounds around which happens a lot when scrolling through tons of sounds on mobile.) For this reason, if you don't understand the code, maybe don't mess with it. Debugging becomes really confusing with all this fragile code.